Datenbank exportieren und Importieren eingebaut, bzw verbessert.
git-svn-id: https://svn.hoffer.cx/repos/redeye/FrameWork/Main/trunk@535 0809fef9-18a3-0410-bae6-c9758b01f9bc

+++ at/redeye/FrameWork/utilities/DeleteDir.java
@@ -16,6 +16,10 @@ public class DeleteDir {
     static public boolean deleteDirectory(File path) {
         if (path.exists()) {
             File[] files = path.listFiles();
+            
+            if( files == null )
+                return true;
+
             for (File file : files) {
                 if (file.isDirectory()) {
                     deleteDirectory(file);
